Remove Result wrapper
Because we are depending on the cross-canister calls to fail, and
because we aren't attempting to coerce the reject code or anything, then
wrapping everything in a Result type seems excessive.

Additionally Dfinity's JS Agent throws if it encounters an err, so we'll
just let those bubble up in our tests rather than explicitly trying to
handle them.