WIP: Start adding `ic.result`
This doesn't work because we `ic.arg_data` is broken in callbacks. It
currently tries to coerce the data into the types of the args of the
method in which it was called, when in reality it should be the type
returned from the cross-canister call.

This code shows a working work-around that works when you call:
- `dfx canister call rejections getResult '(variant {Accept}, "hello")'`
- `dfx canister call rejections getResult '(variant {Reject}, "hello")'`

Additionally, there is commented out code that shows what the API should
look like, but it breaks when passing "accept" because it can't coerce
the boolean that is returned from the "some_service" canister, into the
Variant type of the arg in the "getResult" method.

We need to go back and get `arg_data` working in callbacks before we can
proceed.

Lastly, the tests are included as an example of how it should work, but
they haven't actually been run so there may be errors in there.