Switch benchmarks to a single run
Because the args to each method are always the same, running the
`delete_hero` method over and over again will result in mostly deleting
nonexistent heroes, which isn't what we want for the last two scenarios.
So, to more clearly illustrate the difference between the different
deletion scenarios, we're only doing a single run for now.