Move build failure message to stdout
Stderr is reserved for diganostic information. If you were to redirect
stdout to a file before this change it would say, building canister,
list steps it was going through, and then never print the result if it
failed.

With the change stdout ends with, "build failed" which tells users that
it didn't succeed.

Users can then go see the specifics of why the build didn't succeed over
in stderr which explains the error and gives suggestions on how to fix
it.