🔖 Slim down annotations
Annotations were getting much to complex. This cuts way back on the
complexity, making the error messages much easier to build, at the
expense of making the errors a little less clear.

Given that most of this should move into Rust in the future anyway, this
is a good tradeoff for now.

Note: If we want more complexity in the future, coming back to before
this commit would provide a great starting point.