Mark variables as user-defined
Kybra was already set up to manipulate the variable names outside of the
CDK Framework. This mimics that same pattern, modifying the name before
passing into the CDK Framework.

I do wonder, however, if we should instead have the CDK Framework take
care of this work. The one downside is it might be hard to show/
configure what the variables are renamed to.

Additionally, this prevents issues with variables that clash with rust
keywords.