Streamline null/void handling
I'm not sure if this is more clear, but it's shorter. The main idea
though is that if the return value  doesn't match the return type then
a TypeError should be thrown.

TryFromVmValue for null and undefined both yield the same result
because (()) is the same as (). So, we handle these two cases specially.